# Fix/7358 stock api performance improvement