… tests/api/products-lowinstock.php
… tests/api/products.php