@louwie17 2021-07-19
@moon0326 nice work on this! This is testing well for me, I assume if this goes through you would remove the original `low_in_stock` conditionals within the `Products.php` file?
Also should we still index `variation_id` as part of this PR, or not, given we aren't using it in these queries anymore.
cc: @jeffstieler on this.