@moon0326 2021-07-19
> @moon0326 nice work on this! This is testing well for me, I assume if this goes through you would remove the original `low_in_stock` conditionals within the `Products.php` file?
> Also should we still index `variation_id` as part of this PR, or not, given we aren't using it in these queries anymore.
> cc: @jeffstieler on this.

Yup! Removed `low_in_stock` conditions in https://github.com/woocommerce/woocommerce-admin/pull/7377/commits/39c7ae62e6dafaf7999470fe2842070e57a8e97b

I don't think we need an index for `variation_id` for this PR, but we should go over other API endpoints if we can benefit from the index in a separate issue IMO.

I made a few more updates (https://github.com/woocommerce/woocommerce-admin/pull/7377/commits/bc63d54df5b1a4061c85f3d34b737cfe5ad98e2c, https://github.com/woocommerce/woocommerce-admin/pull/7377/commits/ef4c21bdc968f6ebb204b24d2d92cdcbaa933245). You no longer need to update `Products.php` for testing.