@adrianduffell 2021-07-20
Hey @moon0326 , thanks so much for taking this issue.

The code LGTM and I think this is close 🎉  I think adding the dedicated endpoint was a good choice to allow this query to be standalone and fine-tuned for its purpose.

What do you think of adding the same cache header to this endpoint as in as PR #7364?

```
$response->header( 'Cache-Control', 'max-age=300' );
```

I don't think it’s needed to solve the performance problem, but might be a nice general improvement. I think this data is fetched on every homescreen load at present?

> Yup! Removed low_in_stock conditions in 39c7ae6

We might need to keep this for the time being and deprecate it. I'm not sure if the mobile client depends on it? I'll find out.

I'm going to do further testing today, comparing it to the old query’s results.