@moon0326 2021-07-20
> Hey @moon0326 , thanks so much for taking this issue.
> 
> The code LGTM and I think this is close 🎉 I think adding the dedicated endpoint was a good choice to allow this query to be standalone and fine-tuned for its purpose.
> 
> What do you think of adding the same cache header to this endpoint as in as PR #7364?
> 
> ```
> $response->header( 'Cache-Control', 'max-age=300' );
> ```
> 
> I don't think it’s needed to solve the performance problem, but might be a nice general improvement. I think this data is fetched on every homescreen load at present?
> 
> > Yup! Removed low_in_stock conditions in [39c7ae6](https://github.com/woocommerce/woocommerce-admin/commit/39c7ae62e6dafaf7999470fe2842070e57a8e97b)
> 
> We might need to keep this for the time being and deprecate it. I'm not sure if the mobile client depends on it? I'll find out.
> 
> I'm going to do further testing today, comparing it to the old query’s results.

Thank you!

That's a good idea. I'll add the header.

What do you think about adding the following code to the existing `Products.php`? This should provide the same performance while keeping backward compatibility for those who are already using the endpoint.

Products.php -> get_items();
```
if ($request->get_param('low_in_stock') === true) {
	return (new ProductsLowInStock)->get_items($request);
}
```