@adrianduffell 2021-07-20
> What do you think about adding the following code to the existing `Products.php`?

👍 Nice, that looks like a great approach!

It could still be deprecated by adding a note to the parameter description. For example like what was done here:

https://github.com/woocommerce/woocommerce-admin/blob/38fbb4aedd0f21cf21ddaa6af93e48564fd35324/src/API/Reports/Orders/Stats/Controller.php#L502-L510