@moon0326 2021-07-20
@adrianduffell updated changes in the following commits.

Added deprecated comment: https://github.com/woocommerce/woocommerce-admin/pull/7377/commits/a97c8f536db54560642bfb5c44918486c6bd054a

Use the new controller when `low_in_stock` is set: https://github.com/woocommerce/woocommerce-admin/pull/7377/commits/f203979dd3eadeedc206799418c1711e24689104

Set cache-control: https://github.com/woocommerce/woocommerce-admin/pull/7377/commits/bd33e8c4aabea8bf40c81247ae2d980213deeeea