@adrianduffell 2021-07-20
Awesome, thanks @moon0326 

It looks like the cache header broke a test. I'm investigating.

On a related note, we probably need new PHP tests for the endpoint. In theory it could be a copy/paste of what's there for the deprecated endpoint?