@moon0326 2021-07-20
> Awesome, thanks @moon0326
> 
> It looks like the cache header broke a test. I'm investigating.
> 
> On a related note, we probably need new PHP tests for the endpoint. In theory it could be a copy/paste of what's there for the deprecated endpoint?

Thank you! 

Yeah, I think we can extract out low stock related methods to a new test. I can work on it.