@moon0326 2021-07-20
> Awesome, thanks @moon0326
> 
> It looks like the cache header broke a test. I'm investigating.
> 
> On a related note, we probably need new PHP tests for the endpoint. In theory it could be a copy/paste of what's there for the deprecated endpoint?

Fixed the test and extracted to a new test file in https://github.com/woocommerce/woocommerce-admin/pull/7377/commits/858c10b35a5473ccb8336e0ae5bacfed6809dd55